datum/requirements.txt
Django>=3.0.6,<3.1
djangorestframework==3.12.1
django-cors-headers==3.6.0
requests==2.25.1
uWSGI>=2.0.18,<2.1
redis==3.5.3
rq==1.7.0
mysqlclient==2.0.1
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
